Re-arrange test case
