+++ server/service/src/sync/translations/name.rs
@@ -1,6 +1,6 @@
 use crate::sync::sync_serde::{
     date_option_to_isostring, empty_str_as_option, empty_str_as_option_string,
-    object_fields_as_option, zero_date_as_option,
+     zero_date_as_option,
 };
 use anyhow::Context;
 use chrono::{NaiveDate, NaiveDateTime};
