+++ server/service/src/sync/translations/special/mod.rs
@@ -2,4 +2,3 @@ pub(crate) mod clinician_merge;
 pub(crate) mod item_merge;
 pub(crate) mod name_merge;
 pub(crate) mod name_to_name_store_join;
-pub(crate) mod oms_object_fields;
